refactor: reorganize parser utilities
