+++ services/corrections/item.ts
@@ -4,7 +4,7 @@
  */
 import { Item, AdventureTheme, ItemChange } from '../../types';
 import { MAX_RETRIES, VALID_ITEM_TYPES_STRING } from '../../constants';
-import { isValidItem } from '../validationUtils';
+import { isValidItem } from '../parsers/validation';
 import { formatKnownPlacesForPrompt } from '../../utils/promptFormatters';
 import { callCorrectionAI, callMinimalCorrectionAI } from './base';
